+++ Classification/SGD/src/main/java/org/tribuo/classification/sgd/linear/LinearSGDModel.java
@@ -19,6 +19,7 @@ package org.tribuo.classification.sgd.linear;
 import org.tribuo.Example;
 import org.tribuo.ImmutableFeatureMap;
 import org.tribuo.ImmutableOutputInfo;
+import org.tribuo.ONNXExportable;
 import org.tribuo.Prediction;
 import org.tribuo.classification.Label;
 import org.tribuo.common.sgd.AbstractLinearSGDModel;
@@ -26,9 +27,8 @@ import org.tribuo.math.LinearParameters;
 import org.tribuo.math.la.DenseMatrix;
 import org.tribuo.math.la.DenseVector;
 import org.tribuo.math.util.VectorNormalizer;
-import org.tribuo.onnx.ONNXExportable;
-import org.tribuo.onnx.ONNXNode;
 import org.tribuo.provenance.ModelProvenance;
+import org.tribuo.util.onnx.ONNXNode;
 
 import java.io.IOException;
 import java.util.LinkedHashMap;
