+++ Common/SGD/src/main/java/org/tribuo/common/sgd/AbstractFMModel.java
@@ -22,6 +22,7 @@ import org.tribuo.Example;
 import org.tribuo.Excuse;
 import org.tribuo.ImmutableFeatureMap;
 import org.tribuo.ImmutableOutputInfo;
+import org.tribuo.ONNXExportable;
 import org.tribuo.Output;
 import org.tribuo.math.la.DenseMatrix;
 import org.tribuo.math.la.DenseVector;
@@ -29,13 +30,13 @@ import org.tribuo.math.la.Matrix;
 import org.tribuo.math.la.SGDVector;
 import org.tribuo.math.la.Tensor;
 import org.tribuo.math.onnx.ONNXMathUtils;
-import org.tribuo.onnx.ONNXContext;
-import org.tribuo.onnx.ONNXNode;
-import org.tribuo.onnx.ONNXOperators;
-import org.tribuo.onnx.ONNXPlaceholder;
-import org.tribuo.onnx.ONNXRef;
-import org.tribuo.onnx.ONNXInitializer;
 import org.tribuo.provenance.ModelProvenance;
+import org.tribuo.util.onnx.ONNXContext;
+import org.tribuo.util.onnx.ONNXInitializer;
+import org.tribuo.util.onnx.ONNXNode;
+import org.tribuo.util.onnx.ONNXOperators;
+import org.tribuo.util.onnx.ONNXPlaceholder;
+import org.tribuo.util.onnx.ONNXRef;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -267,7 +268,7 @@ public abstract class AbstractFMModel<T extends Output<T>> extends AbstractSGDMo
         ONNXPlaceholder input = onnx.floatInput("input", featureIDMap.size());
         ONNXPlaceholder output = onnx.floatOutput("output", outputIDInfo.size());
         writeONNXGraph(input).assignTo(output);
-        return onnx.model(domain, modelVersion, this);
+        return ONNXExportable.buildModel(onnx, domain, modelVersion, this);
     }
 
 }
