+++ Common/SGD/src/main/java/org/tribuo/common/sgd/AbstractLinearSGDModel.java
@@ -24,18 +24,19 @@ import org.tribuo.Feature;
 import org.tribuo.ImmutableFeatureMap;
 import org.tribuo.ImmutableOutputInfo;
 import org.tribuo.Model;
+import org.tribuo.ONNXExportable;
 import org.tribuo.Output;
 import org.tribuo.Prediction;
 import org.tribuo.math.LinearParameters;
 import org.tribuo.math.la.DenseMatrix;
 import org.tribuo.math.la.Matrix;
-import org.tribuo.onnx.ONNXContext;
-import org.tribuo.onnx.ONNXNode;
-import org.tribuo.onnx.ONNXOperators;
-import org.tribuo.onnx.ONNXPlaceholder;
-import org.tribuo.onnx.ONNXRef;
-import org.tribuo.onnx.ONNXInitializer;
 import org.tribuo.provenance.ModelProvenance;
+import org.tribuo.util.onnx.ONNXContext;
+import org.tribuo.util.onnx.ONNXInitializer;
+import org.tribuo.util.onnx.ONNXNode;
+import org.tribuo.util.onnx.ONNXOperators;
+import org.tribuo.util.onnx.ONNXPlaceholder;
+import org.tribuo.util.onnx.ONNXRef;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -213,7 +214,7 @@ public abstract class AbstractLinearSGDModel<T extends Output<T>> extends Abstra
         ONNXPlaceholder input = onnx.floatInput("input", featureIDMap.size());
         ONNXPlaceholder output = onnx.floatOutput("output", outputIDInfo.size());
         writeONNXGraph(input).assignTo(output);
-        return onnx.model(domain, modelVersion, this);
+        return ONNXExportable.buildModel(onnx, domain, modelVersion, this);
     }
 
 }
