+++ Math/src/main/java/org/tribuo/math/util/SigmoidNormalizer.java
@@ -16,8 +16,8 @@
 
 package org.tribuo.math.util;
 
-import org.tribuo.onnx.ONNXNode;
-import org.tribuo.onnx.ONNXOperators;
+import org.tribuo.util.onnx.ONNXNode;
+import org.tribuo.util.onnx.ONNXOperators;
 
 import java.io.Serializable;
 import java.util.Arrays;
