+++ MultiLabel/SGD/src/main/java/org/tribuo/multilabel/sgd/fm/FMMultiLabelModel.java
@@ -19,6 +19,7 @@ package org.tribuo.multilabel.sgd.fm;
 import org.tribuo.Example;
 import org.tribuo.ImmutableFeatureMap;
 import org.tribuo.ImmutableOutputInfo;
+import org.tribuo.ONNXExportable;
 import org.tribuo.Prediction;
 import org.tribuo.classification.Label;
 import org.tribuo.common.sgd.AbstractFMModel;
@@ -26,9 +27,8 @@ import org.tribuo.common.sgd.FMParameters;
 import org.tribuo.math.la.DenseVector;
 import org.tribuo.math.util.VectorNormalizer;
 import org.tribuo.multilabel.MultiLabel;
-import org.tribuo.onnx.ONNXExportable;
-import org.tribuo.onnx.ONNXNode;
 import org.tribuo.provenance.ModelProvenance;
+import org.tribuo.util.onnx.ONNXNode;
 
 import java.util.HashMap;
 import java.util.HashSet;
