+++ Util/ONNXExport/src/main/java/org/tribuo/util/onnx/ONNXAttribute.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.tribuo.onnx;
+package org.tribuo.util.onnx;
 
 import com.google.protobuf.ByteString;
 import ai.onnx.proto.OnnxMl;
