+++ Util/ONNXExport/src/main/java/org/tribuo/util/onnx/ONNXInitializer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.tribuo.onnx;
+package org.tribuo.util.onnx;
 
 import ai.onnx.proto.OnnxMl;
