+++ Util/ONNXExport/src/main/java/org/tribuo/util/onnx/package-info.java
@@ -15,13 +15,9 @@
  */
 
 /**
- * Interfaces and utilities for exporting Tribuo {@link org.tribuo.Model}s in
- * <a href="https://onnx.ai">ONNX</a> format.
+ * Interfaces and utilities for writing <a href="https://onnx.ai>ONNX</a> models from Java.
  * <p>
- * ONNX exported models use floats where Tribuo uses doubles, this is due
- * to comparatively poor support for fp64 in ONNX deployment environments
- * as compared to fp32. In addition fp32 executes better on the various
- * accelerator backends available in
- * <a href="https://onnxruntime.ai">ONNX Runtime</a>.
+ * Developed to support <a href="https://tribuo.org">Tribuo</a>, but can be used to export
+ * other machine learning models from JVM languages.
  */
-package org.tribuo.onnx;
+package org.tribuo.util.onnx;
