+++ Util/ONNXExport/src/main/resources/LICENSE.txt
@@ -1,0 +1,1 @@
+../../../../../LICENSE.txt
