+++ Util/ONNXExport/src/main/resources/THIRD_PARTY_LICENSES.txt
@@ -1,0 +1,1 @@
+../../../../../THIRD_PARTY_LICENSES.txt
